Formatierungen ausgelagert
